--- a/src/api/data/fixtures/readings/2014-06-09-01-shootout-at-joes-by-richard-laymon.md
+++ b/src/api/data/fixtures/readings/2014-06-09-01-shootout-at-joes-by-richard-laymon.md
@@ -1,9 +1,0 @@
----
-sequence: 1
-work_slug: shootout-at-joes-by-richard-laymon
-edition: Paperback
-edition_notes: Globe Fearon Co, 1984
-timeline:
-  - date: 2014-06-09
-    progress: Finished
----