--- a/src/api/data/fixtures/readings/2014-06-09-02-dawsons-city-by-richard-laymon.md
+++ b/src/api/data/fixtures/readings/2014-06-09-02-dawsons-city-by-richard-laymon.md
@@ -1,9 +1,0 @@
----
-sequence: 2
-work_slug: dawsons-city-by-richard-laymon
-edition: Paperback
-edition_notes: Globe Fearon Co, 1984
-timeline:
-  - date: 2014-06-09
-    progress: Finished
----