--- a/src/api/data/fixtures/readings/2014-07-13-01-the-last-hand-by-richard-laymon.md
+++ b/src/api/data/fixtures/readings/2014-07-13-01-the-last-hand-by-richard-laymon.md
@@ -1,9 +1,0 @@
----
-sequence: 1
-work_slug: the-last-hand-by-richard-laymon
-edition: Paperback
-edition_notes: Lake Pub Co, 1985
-timeline:
-  - date: 2014-07-13
-    progress: Finished
----