--- a/src/api/data/fixtures/readings/2014-11-02-01-live-bait-by-richard-laymon.md
+++ b/src/api/data/fixtures/readings/2014-11-02-01-live-bait-by-richard-laymon.md
@@ -1,9 +1,0 @@
----
-sequence: 1
-work_slug: live-bait-by-richard-laymon
-edition: Paperback
-edition_notes: Fearon, 1985
-timeline:
-  - date: 2014-11-02
-    progress: Finished
----