--- a/src/api/data/fixtures/readings/2017-12-05-01-shopgirl-by-steve-martin.md
+++ b/src/api/data/fixtures/readings/2017-12-05-01-shopgirl-by-steve-martin.md
@@ -1,9 +1,0 @@
----
-sequence: 1
-work_slug: shopgirl-by-steve-martin
-edition: Audible
-edition_notes: null
-timeline:
-  - date: 2017-12-05
-    progress: Finished
----