--- a/src/api/data/fixtures/readings/2022-09-11-01-draft-no-4-by-john-mcphee.md
+++ b/src/api/data/fixtures/readings/2022-09-11-01-draft-no-4-by-john-mcphee.md
@@ -1,9 +1,0 @@
----
-sequence: 1
-work_slug: draft-no-4-by-john-mcphee
-edition: Audible
-edition_notes: Read by John McPhee
-timeline:
-  - date: 2022-09-11
-    progress: Finished
----