--- a/src/api/data/fixtures/readings/2023-10-22-01-the-creative-act-by-rick-rubin.md
+++ b/src/api/data/fixtures/readings/2023-10-22-01-the-creative-act-by-rick-rubin.md
@@ -1,9 +1,0 @@
----
-sequence: 1
-work_slug: the-creative-act-by-rick-rubin
-edition: Audible
-edition_notes: Read by Rick Rubin
-timeline:
-  - date: 2023-10-22
-    progress: Finished
----