--- a/src/api/data/fixtures/reviews/brain-wave-by-poul-anderson.md
+++ b/src/api/data/fixtures/reviews/brain-wave-by-poul-anderson.md
@@ -1,9 +1,0 @@
----
-work_slug: brain-wave-by-poul-anderson
-grade: C
-date: 2013-04-07
----
-
-Humanity struggles to adapt after a cosmic event elevates the intelligence of every thinking creature on Earth, turning simpletons into geniuses, and average men and women into near gods.
-
-Anderson uses the ripe premise to explore the social implications of intelligence, from the way man interacts with lesser species to the economic structures that make our lives possible. Though some of Anderson’s conclusions betray the book’s near half-century age (the notion of cognitive bias wouldn’t be introduced for nearly 15 years), it’s still an insightful, well-written (at least the first and third acts) commentary that ages remarkably well.