--- a/src/api/data/fixtures/reviews/cary-grant-by-marc-eliot.md
+++ b/src/api/data/fixtures/reviews/cary-grant-by-marc-eliot.md
@@ -1,9 +1,0 @@
----
-work_slug: cary-grant-by-marc-eliot
-grade: C
-date: 2012-05-31
----
-
-While it's clear Elliot (like most of the world) harbors a love for Grant's cinematic persona, his well sourced biography won't be for everyone, as it makes a case for Grant's long-rumored bi-sexuality and paints him as a man haunted by insecurity and incapable of living up to his own image.
-
-Those looking to maintain the illusion that Grant the character and Grant the man were one and the same would do better to read Grant's autobiography, but for those looking for a peek behind the curtain, this is a brisk, entertaining read.