--- a/src/api/data/fixtures/reviews/dark-crusade-by-karl-edward-wagner.md
+++ b/src/api/data/fixtures/reviews/dark-crusade-by-karl-edward-wagner.md
@@ -1,9 +1,0 @@
----
-work_slug: dark-crusade-by-karl-edward-wagner
-grade: C
-date: 2012-05-18
----
-
-After a botched power-play leaves him a fugitive, Wagner's anti-hero Kane allies himself with Orted Ak-Ceddi, a former bandit now possessed by a forgotten god, whose cult has swarmed in numbers to conquer cities and even kingdoms before finally suffering their first defeat. Kane transforms Ak-Ceddi's untrained mob into an army and the bloody conquests resume, though both men know it's only matter of time before they will clash for control of the power they wield.
-
-Wagner's nihilistic tale sees an immoral prophet clash with an amoral immortal, and it's left to reader to decide which is the true evil. While it's certainly nice to see someone so ably pick up where Robert E. Howard left off, Wagner assumes a knowledge of military terminology that makes many of his larger set pieces confusing. That, combined with the story's fragmented nature makes this something of a frustrating, though enjoyable, read.