--- a/src/api/data/fixtures/reviews/deliverance-by-james-dickey.md
+++ b/src/api/data/fixtures/reviews/deliverance-by-james-dickey.md
@@ -1,6 +1,0 @@
----
-work_slug: deliverance-by-james-dickey
-grade: B
-date: 2017-11-12
----
-