--- a/src/api/data/fixtures/reviews/is-this-anything-by-jerry-seinfeld.md
+++ b/src/api/data/fixtures/reviews/is-this-anything-by-jerry-seinfeld.md
@@ -1,9 +1,0 @@
----
-work_slug: is-this-anything-by-jerry-seinfeld
-grade: C
-date: 2022-09-10
----
-
-Comedian Jerry Seinfeld’s standup material arranged in chronological order with autobiographical interludes. It begins with his first successful bits—one of my favorites concerns Life cereal—builds to his star making _Tonight Show_ appearance and carries through his self-titled sitcom and _I’m Telling You for the Last Time_ HBO special.
-
-At this point, Seinfeld had retired. But seeing Chris Rock perform reignited his passion for the art. Rather than recycle his old material, he starts over. The later third comprises bits he developed in the New York clubs post-stardom. It has a distinct voice, one older and crankier, but still Seinfeld. While amusing, these bits lacked instant classics like “The Helmet” or “Public Speaking.”