--- a/src/api/data/fixtures/reviews/midnights-lair-by-richard-laymon.md
+++ b/src/api/data/fixtures/reviews/midnights-lair-by-richard-laymon.md
@@ -1,6 +1,0 @@
----
-work_slug: midnights-lair-by-richard-laymon
-grade: C
-date: 2018-01-11
----
-