--- a/src/api/data/fixtures/reviews/mr-s-by-george-jacobs-william-stadiem.md
+++ b/src/api/data/fixtures/reviews/mr-s-by-george-jacobs-william-stadiem.md
@@ -1,9 +1,0 @@
----
-work_slug: mr-s-by-george-jacobs-william-stadiem
-grade: B
-date: 2012-06-14
----
-
-Jacobs offers an insiders point of view into Sinatra's life beginning just before his comeback performance in _From Here to Eternity_ and ending in the aftermath of his troubled marriage to Mia Farrow.
-
-While certainly salacious, the love Jacob's harbors for Mr. S. is evident, resulting in a raconteur tale that may not be the gospel truth, but is one hell of an enjoyable read.