--- a/src/api/data/fixtures/reviews/the-48-laws-of-power-by-robert-greene.md
+++ b/src/api/data/fixtures/reviews/the-48-laws-of-power-by-robert-greene.md
@@ -1,7 +1,0 @@
----
-work_slug: the-48-laws-of-power-by-robert-greene
-grade: D+
-date: 2022-07-15
----
-
-A passable synthesis of Sun Tzu’s <span data-work-slug="the-art-of-war-by-sun-tzu">_The Art of War_</span>, Machiavelli's <span data-work-slug="the-prince-by-niccolo-machiavelli">_The Prince_</span>, and Carl von Clausewitz’s <span data-work-slug="on-war-by-carl-von-clausewitz">_On War_</span> that reads somewhere between the source texts and CliffsNotes. Greene presents the book in an episodic manner. Laws don’t build on one-another and he re-introduces historical figures like Mata Hari and Cardinal Richelieu each time he cites them as examples. Written prior to social media’s emergence and claiming Henry Kissinger as a contemporary example, it—unlike its source text—feels dated. An updated version leading with modern examples and working back to the fundamental laws might fare better.