--- a/src/api/data/fixtures/reviews/the-long-walk-by-stephen-king.md
+++ b/src/api/data/fixtures/reviews/the-long-walk-by-stephen-king.md
@@ -1,9 +1,0 @@
----
-work_slug: the-long-walk-by-stephen-king
-grade: D
-date: 2014-03-09
----
-
-_The Long Walk_ is set in a dystopian present. The United States is under the control of a totalitarian government. Once a year, the country's attention turns to 'The Long Walk,' a walking contest composed of one hundred teenaged boys. The competition starts in Main and continues down the eastern seaboard until a single walker remains. No stops, no breaks, and if you drop below a pace of four miles per hour for more than thirty seconds you get a warning. Three warnings and you're shot dead.
-
-The story is less about the walk and more about the mental state of the characters walking. With nothing to do but walk, the characters talk in long soliloquies. It's a clear allegory to the Vietnam War, which was in full swing when the then-18-year-old King was writing the book. To that end, it's a an effective timecapsule of a generation's angst and cynicism. But as a novel, it's a bit more problematic. Many of the characters bleed into one-another, speaking with the same voice, only espousing different viewpoints. It's as if King was more concerned with what he wanted to say about the war than with the story.