… src/api/data/fixtures/data/authors/a-jack-ulrich.json
… src/api/data/fixtures/data/authors/arnold-schwarzenegger.json
… src/api/data/fixtures/data/authors/brad-schoenfeld.json
… src/api/data/fixtures/data/authors/bret-easton-ellis.json
… src/api/data/fixtures/data/authors/conner-habib.json
… src/api/data/fixtures/data/authors/dave-draper.json
… src/api/data/fixtures/data/authors/david-mamet.json
… src/api/data/fixtures/data/authors/david-miller.json
… src/api/data/fixtures/data/authors/donna-tartt.json
… src/api/data/fixtures/data/authors/edward-lee.json
… src/api/data/fixtures/data/authors/george-jacobs.json
… src/api/data/fixtures/data/authors/graham-masterton.json
… src/api/data/fixtures/data/authors/james-m-cain.json
… src/api/data/fixtures/data/authors/jerry-seinfeld.json
… src/api/data/fixtures/data/authors/joe-kenda.json
… src/api/data/fixtures/data/authors/john-mcphee.json
… src/api/data/fixtures/data/authors/johnny-marr.json
… src/api/data/fixtures/data/authors/karl-edward-wagner.json
… src/api/data/fixtures/data/authors/kathe-koja.json
… src/api/data/fixtures/data/authors/leslie-h-whitten-jr.json
… src/api/data/fixtures/data/authors/lol-tolhurst.json
… src/api/data/fixtures/data/authors/marc-eliot.json
… src/api/data/fixtures/data/authors/nigel-andrews.json
… src/api/data/fixtures/data/authors/peter-straub.json
… src/api/data/fixtures/data/authors/poul-anderson.json
… src/api/data/fixtures/data/authors/raymond-chandler.json
… src/api/data/fixtures/data/authors/rick-rubin.json
… src/api/data/fixtures/data/authors/robert-greene.json
… src/api/data/fixtures/data/authors/ross-macdonald.json
… src/api/data/fixtures/data/authors/scott-carney.json