## Summary
- Reduced test fixtures from 82 to 27 reviews while maintaining code coverage
- Test runtime improved by ~32% (12.48s → 8.54s)
- Maintained all foreign key relationships between fixtures

## Changes
- **Pruned test fixtures** to minimal set covering all test permutations:
  - Reviews: 82 → 27 (67% reduction)
  - Authors: 52 → 19 (63% reduction)
  - Readings: 201 → 54 (73% reduction)
  - Timeline entries: 899 → 181 (80% reduction)

- **Updated test files** to use available years in reduced dataset
- **Regenerated all snapshots** to reflect smaller fixture set

## Test Coverage
- Maintained 91.86% overall code coverage
- All 165 tests passing
- Coverage includes:
  - All 5 work kinds (Novel, Nonfiction, Collection, Novella, Short Story)
  - 11 of 13 grade variations
  - Special cases (multiple authors, subtitles, audiobooks, abandoned readings)

## Performance Impact
Before: 12.48s test runtime
After: 8.54s test runtime
**Improvement: ~32% faster**

🤖 Generated with [Claude Code](https://claude.ai/code)